+++ src/FloatTypeArray.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Fortuneglobe\Types;
+
+use Fortuneglobe\Types\Interfaces\RepresentsFloatType;
+
+class FloatTypeArray extends AbstractFloatTypeArray
+{
+	public static function isValid( RepresentsFloatType $floatType ): bool
+	{
+		return true;
+	}
+}